Add wrapper protocol for all existing functions
Up until now the package only provided functions. The issue with this
approach is you would need to provide your own interfaces for the
functions and inject them as dependencies if you wanted to be able to
unit test their consumers.

With the introduction of the `JsonApiCompatible` protocol and the
default implementation via the protocol extension, it should now be
possible to mock selected methods for easier unit testing.